Removing integration tests relying on an already running instance of eventstore in favour of doicker automated ones used for both ESJC and the new gRPC driver
