+++ sourcerer-eventstore-test-utils/build.gradle
@@ -1,0 +1,14 @@
+apply plugin: "kotlin"
+
+dependencies {
+    api project(':sourcerer-core')
+    implementation 'junit:junit:4.12'
+    implementation 'org.hamcrest:hamcrest-all:1.3'
+    implementation 'org.mockito:mockito-core:1.10.19'
+    implementation platform('io.projectreactor:reactor-bom:2020.0.18')
+    implementation 'io.projectreactor:reactor-core'
+    implementation 'com.fasterxml.jackson.core:jackson-databind:2.13.2.2'
+    implementation 'com.github.docker-java:docker-java-core:3.2.13'
+    implementation 'com.github.docker-java:docker-java-transport-httpclient5:3.2.13'
+    implementation 'com.fasterxml.jackson.module:jackson-module-kotlin:2.13.2'
+}
